[refactor] book 엔티티 bookNo 필드 타입 변경에 따른 코드 수정

+++ src/main/java/com/team/student_calendar/dto/BookCreateReq.java
@@ -23,7 +23,7 @@ public class BookCreateReq {
 
     private Integer difficulty;
 
-    private Integer bookNo;
+    private Long bookNo;
 
     private String imageUrl;
